requirements.txt
antispamscript==0.0.1
asn1crypto==0.24.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
http-ece==1.1.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
mailapi==0.0.6
MarkupSafe==1.1.1
pkginfo==1.5.0.1
py-vapid==1.7.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pywebpush==1.10.1
PyYAML==5.1.2
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
schedule==0.6.0
six==1.12.0
tqdm==4.32.2
twine==1.13.0
urllib3==1.25.3
webencodings==0.5.1
Werkzeug==0.15.5
